refactor(lunch-service): use const to get random description and title
