+++ src/features/discord/messages/enums/commands/lunch/discord-message-command-lunch-description.enum.ts
@@ -1,5 +1,6 @@
 export enum DiscordMessageCommandLunchDescriptionEnum {
   AWESOME = `Awesome!`,
+  COOL = `Cool.`,
   FOOD_IS_HEART = `Food is :heart:`,
   FOOD_IS_HEARTS = `Food is :hearts:`,
   FOOD_IS_YUM = `Food is :yum:`,
