+++ src/features/discord/messages/services/command/lunch/services/discord-message-command-lunch.service.ts
@@ -6,18 +6,17 @@ import {
 } from "discord.js";
 import _ from "lodash";
 import moment from "moment-timezone";
-import { AbstractService } from "../../../../../../classes/services/abstract.service";
-import { ServiceNameEnum } from "../../../../../../enums/service-name.enum";
-import { getRandomValueFromEnum } from "../../../../../../functions/randoms/get-random-value-from-enum";
-import { LoggerService } from "../../../../../logger/services/logger.service";
-import { DiscordSoniaService } from "../../../../users/services/discord-sonia.service";
-import { DiscordMessageCommandEnum } from "../../../enums/commands/discord-message-command.enum";
-import { DiscordMessageCommandLunchDescriptionEnum } from "../../../enums/commands/lunch/discord-message-command-lunch-description.enum";
-import { DiscordMessageCommandLunchTitleEnum } from "../../../enums/commands/lunch/discord-message-command-lunch-title.enum";
-import { discordHasThisCommand } from "../../../functions/commands/checks/discord-has-this-command";
-import { IDiscordMessageResponse } from "../../../interfaces/discord-message-response";
-import { IAnyDiscordMessage } from "../../../types/any-discord-message";
-import { DiscordMessageConfigService } from "../../config/discord-message-config.service";
+import { AbstractService } from "../../../../../../../classes/services/abstract.service";
+import { ServiceNameEnum } from "../../../../../../../enums/service-name.enum";
+import { LoggerService } from "../../../../../../logger/services/logger.service";
+import { DiscordSoniaService } from "../../../../../users/services/discord-sonia.service";
+import { DiscordMessageCommandEnum } from "../../../../enums/commands/discord-message-command.enum";
+import { discordHasThisCommand } from "../../../../functions/commands/checks/discord-has-this-command";
+import { IDiscordMessageResponse } from "../../../../interfaces/discord-message-response";
+import { IAnyDiscordMessage } from "../../../../types/any-discord-message";
+import { DiscordMessageConfigService } from "../../../config/discord-message-config.service";
+import { DISCORD_MESSAGE_COMMAND_LUNCH_DESCRIPTION_MESSAGES } from "../constants/discord-message-command-lunch-description-messages";
+import { DISCORD_MESSAGE_COMMAND_LUNCH_TITLE_MESSAGES } from "../constants/discord-message-command-lunch-title-messages";
 
 export class DiscordMessageCommandLunchService extends AbstractService {
   private static _instance: DiscordMessageCommandLunchService;
@@ -88,10 +87,7 @@ export class DiscordMessageCommandLunchService extends AbstractService {
   }
 
   private _getMessageDescription(): string {
-    return (
-      getRandomValueFromEnum(DiscordMessageCommandLunchDescriptionEnum) ||
-      `Cool.`
-    );
+    return DISCORD_MESSAGE_COMMAND_LUNCH_DESCRIPTION_MESSAGES.getRandomMessage();
   }
 
   private _getMessageEmbedFooter(): MessageEmbedFooter {
@@ -116,9 +112,6 @@ export class DiscordMessageCommandLunchService extends AbstractService {
   }
 
   private _getMessageEmbedTitle(): string {
-    return (
-      getRandomValueFromEnum(DiscordMessageCommandLunchTitleEnum) ||
-      `Lunch time!`
-    );
+    return DISCORD_MESSAGE_COMMAND_LUNCH_TITLE_MESSAGES.getRandomMessage();
   }
 }
